+++ bluefin/src/Bluefin/Compound.hs
@@ -664,6 +664,7 @@ module Bluefin.Compound
     OneWayCoercible.OneWayCoercible (OneWayCoercible.oneWayCoercibleImpl),
     OneWayCoercibleHandle (MkOneWayCoercibleHandle),
     OneWayCoercible.gOneWayCoercible,
+    oneWayCoercibleTrustMe,
     -- | Bluefin re-exports @Generic@ for convenience.
     OneWayCoercible.Generic,